refactor(tax): unify state tax and state capital gains calculators
